+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 ### Added
+- Add support for retrieving the current values of all stats.
 
 ## [3.1.0] - 2018-04-20
 Initial open source release.